+++ akshare/stock/stock_zh_a_sina.py
@@ -1,5 +1,5 @@
-# -*- coding:utf-8 -*-
 #!/usr/bin/env python
+# -*- coding:utf-8 -*-
 """
 Date: 2021/9/8 19:28
 Desc: 新浪财经-A股-实时行情数据和历史行情数据(包含前复权和后复权因子)
@@ -376,7 +376,7 @@ def stock_zh_a_minute(
 
 
 if __name__ == "__main__":
-    stock_zh_a_daily_hfq_df_one = stock_zh_a_daily(symbol="sh603843", start_date="20171103", end_date="20210908", adjust="")
+    stock_zh_a_daily_hfq_df_one = stock_zh_a_daily(symbol="sh600308", start_date="20171103", end_date="20211021", adjust="qfq")
     print(stock_zh_a_daily_hfq_df_one)
 
     stock_zh_a_daily_hfq_df_three = stock_zh_a_daily(symbol="sz000001", start_date="19900103", end_date="20210118", adjust="qfq")